rename: helper, mapper 클래스 재지정
